+++ cmd/main.go
@@ -354,8 +354,9 @@ func main() {
 		os.Exit(1)
 	}
 	if err := (&controller.HomeAssistantConfigurationReconciler{
-		Client: mgr.GetClient(),
-		Scheme: mgr.GetScheme(),
+		Client:   mgr.GetClient(),
+		Scheme:   mgr.GetScheme(),
+		Recorder: mgr.GetEventRecorder("homeassistantconfiguration-controller"),
 	}).SetupWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create controller", "controller", "HomeAssistantConfiguration")
 		os.Exit(1)